메뉴 DB연결 조정

+++ POS/Entity/MenuEntity.cs
@@ -20,11 +20,11 @@ namespace POS.Entity
 
         [Required]
         [Column("menu_price")]
-        public decimal menuPrice { get; set; }
+        public int menuPrice { get; set; }
 
         public MenuEntity() { }
 
-        public MenuEntity(string name, decimal price)
+        public MenuEntity(string name, int price)
         {
             menuName = name;
             menuPrice = price;
